Merge the v0.2.2 patch development branch.

Changelog

- [bugfix]   Fix a bug that prevented removing slide assets that didn't
             have a thumbnail.
- [bugfix]   Fix dialogs not appearing correctly on the documentation
             pages.
- [bugfix]   Fix a bug where adding media to user-created slides didn't
             work. (#27)
- [feature]  Implement a new tag [font] in the markup transpiler for
             changing fonts in slides. (#28)
- [bugfix]   Add a missing class definition that caused problems when
             uploading files.
- [bugfix]   Fix a bug where the editor media uploader popup didn't
             close if the (X) button was clicked after an upload had
             failed.
- [feature]  Allow space characters in upload filenames.
- [bugfix]   Implement automatic slide lock cleanup on logout.

+++ src/common/php/slide/slideasset.php
@@ -9,7 +9,7 @@ require_once($_SERVER['DOCUMENT_ROOT'].'/common/php/config.php');
 require_once($_SERVER['DOCUMENT_ROOT'].'/common/php/exportable/exportable.php');
 require_once($_SERVER['DOCUMENT_ROOT'].'/common/php/thumbnail/thumbnail.php');
 
-const ASSET_FILENAME_REGEX = '/^[A-Za-z0-9_.-]*$/';
+const ASSET_FILENAME_REGEX = '/^[ A-Za-z0-9_.-]*$/';
 
 class SlideAsset extends Exportable {
 	static $PRIVATE = [
@@ -98,11 +98,19 @@ class SlideAsset extends Exportable {
 	}
 
 	public function remove() {
-		if (!empty($this->fullpath)) {
+		if (
+			!empty($this->fullpath)
+			&& is_file($this->fullpath)
+		) {
 			unlink($this->fullpath);
-			unlink($this->thumbpath);
-			$this->reset();
 		}
+		if (
+			!empty($this->thumbpath)
+			&& is_file($this->thumbpath)
+		) {
+			unlink($this->thumbpath);
+		}
+		$this->reset();
 	}
 
 	private function reset() {
